discard modification on gau
