+++ mmpose/models/utils/transformer.py
@@ -1,16 +1,12 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 import math
-from typing import Optional, Sequence, Union
+from typing import Sequence
 
-import torch
 import torch.nn as nn
 import torch.nn.functional as F
-from mmcv.cnn import Scale, build_conv_layer, build_norm_layer
-from mmcv.cnn.bricks import DropPath
+from mmcv.cnn import build_conv_layer, build_norm_layer
 from mmengine.model import BaseModule
-from mmengine.utils import digit_version, to_2tuple
-from mmengine.utils.dl_utils import TORCH_VERSION
-from torch import Tensor
+from mmengine.utils import to_2tuple
 
 
 def nlc_to_nchw(x, hw_shape):
@@ -371,341 +367,3 @@ class PatchMerging(BaseModule):
         x = self.norm(x) if self.norm else x
         x = self.reduction(x)
         return x, output_size
-
-
-class ScaleNorm(nn.Module):
-    """Scale Norm.
-
-    Args:
-        dim (int): The dimension of the scale vector.
-        eps (float, optional): The minimum value in clamp. Defaults to 1e-5.
-
-    Reference:
-        `Transformers without Tears: Improving the Normalization
-        of Self-Attention <https://arxiv.org/abs/1910.05895>`_
-    """
-
-    def __init__(self, dim, eps=1e-5):
-        super().__init__()
-        self.scale = dim**-0.5
-        self.eps = eps
-        self.g = nn.Parameter(torch.ones(1))
-
-    def forward(self, x):
-        """Forward function.
-
-        Args:
-            x (torch.Tensor): Input tensor.
-
-        Returns:
-            torch.Tensor: The tensor after applying scale norm.
-        """
-
-        norm = torch.norm(x, dim=-1, keepdim=True) * self.scale
-        return x / norm.clamp(min=self.eps) * self.g
-
-
-class SinePositionalEncoding(nn.Module):
-    """Sine Positional Encoding Module. This module implements sine positional
-    encoding, which is commonly used in transformer-based models to add
-    positional information to the input sequences. It uses sine and cosine
-    functions to create positional embeddings for each element in the input
-    sequence.
-
-    Args:
-        out_channels (int): The number of features in the input sequence.
-        temperature (int): A temperature parameter used to scale
-            the positional encodings. Defaults to 10000.
-        spatial_dim (int): The number of spatial dimension of input
-            feature. 1 represents sequence data and 2 represents grid data.
-            Defaults to 1.
-        learnable (bool): Whether to optimize the frequency base. Defaults
-            to False.
-        eval_size (int, tuple[int], optional): The fixed spatial size of
-            input features. Defaults to None.
-    """
-
-    def __init__(
-        self,
-        out_channels: int,
-        spatial_dim: int = 1,
-        temperature: int = 1e5,
-        learnable: bool = False,
-        eval_size: Optional[Union[int, Sequence[int]]] = None,
-    ) -> None:
-
-        super().__init__()
-
-        assert out_channels % 2 == 0
-        assert temperature > 0
-
-        self.spatial_dim = spatial_dim
-        self.out_channels = out_channels
-        self.temperature = temperature
-        self.eval_size = eval_size
-        self.learnable = learnable
-
-        pos_dim = out_channels // 2
-        dim_t = torch.arange(pos_dim, dtype=torch.float32) / pos_dim
-        dim_t = self.temperature**(dim_t)
-
-        if not learnable:
-            self.register_buffer('dim_t', dim_t)
-        else:
-            self.dim_t = nn.Parameter(dim_t, requires_grad=True)
-
-        # reset parameters
-        if eval_size:
-            if hasattr(self, f'pos_enc_{eval_size}'):
-                delattr(self, f'pos_enc_{eval_size}')
-            pos_enc = self.generate_pos_encoding(size=eval_size)
-            self.register_buffer(f'pos_enc_{eval_size}', pos_enc)
-
-    def forward(self, *args, **kwargs):
-        return self.generate_pos_encoding(*args, **kwargs)
-
-    def generate_pos_encoding(self,
-                              size: Union[int, Sequence[int]] = None,
-                              position: Optional[Tensor] = None):
-        """Generate positional encoding for input features.
-
-        Args:
-            size (int or tuple[int]): Size of the input features. Required
-                if position is None.
-            position (Tensor, optional): Position tensor. Required if size
-                is None.
-        """
-
-        assert (size is not None) ^ (position is not None)
-
-        if (not (self.learnable
-                 and self.training)) and size is not None and hasattr(
-                     self, f'pos_enc_{size}'):
-            return getattr(self, f'pos_enc_{size}')
-
-        if self.spatial_dim == 1:
-            if size is not None:
-                if isinstance(size, (tuple, list)):
-                    size = size[0]
-                position = torch.arange(
-                    size, dtype=torch.float32, device=self.dim_t.device)
-
-            dim_t = self.dim_t.reshape(*((1, ) * position.ndim), -1)
-            freq = position.unsqueeze(-1) / dim_t
-            pos_enc = torch.cat((freq.cos(), freq.sin()), dim=-1)
-
-        elif self.spatial_dim == 2:
-            if size is not None:
-                if isinstance(size, (tuple, list)):
-                    h, w = size[:2]
-                elif isinstance(size, (int, float)):
-                    h, w = int(size), int(size)
-                else:
-                    raise ValueError(f'got invalid type {type(size)} for size')
-                grid_h, grid_w = torch.meshgrid(
-                    torch.arange(
-                        int(h), dtype=torch.float32, device=self.dim_t.device),
-                    torch.arange(
-                        int(w), dtype=torch.float32, device=self.dim_t.device))
-                grid_h, grid_w = grid_h.flatten(), grid_w.flatten()
-            else:
-                assert position.size(-1) == 2
-                grid_h, grid_w = torch.unbind(position, dim=-1)
-
-            dim_t = self.dim_t.reshape(*((1, ) * grid_h.ndim), -1)
-            freq_h = grid_h.unsqueeze(-1) / dim_t
-            freq_w = grid_w.unsqueeze(-1) / dim_t
-            pos_enc_h = torch.cat((freq_h.cos(), freq_h.sin()), dim=-1)
-            pos_enc_w = torch.cat((freq_w.cos(), freq_w.sin()), dim=-1)
-            pos_enc = torch.stack((pos_enc_h, pos_enc_w), dim=-1)
-
-        return pos_enc
-
-    @staticmethod
-    def apply_additional_pos_enc(feature: Tensor,
-                                 pos_enc: Tensor,
-                                 spatial_dim: int = 1):
-        """Apply additional positional encoding to input features.
-
-        Args:
-            feature (Tensor): Input feature tensor.
-            pos_enc (Tensor): Positional encoding tensor.
-            spatial_dim (int): Spatial dimension of input features.
-        """
-
-        assert spatial_dim in (1, 2), f'the argument spatial_dim must be ' \
-            f'either 1 or 2, but got {spatial_dim}'
-        if spatial_dim == 2:
-            pos_enc = pos_enc.flatten(-2)
-        for _ in range(feature.ndim - pos_enc.ndim):
-            pos_enc = pos_enc.unsqueeze(0)
-        return feature + pos_enc
-
-    @staticmethod
-    def apply_rotary_pos_enc(feature: Tensor,
-                             pos_enc: Tensor,
-                             spatial_dim: int = 1):
-        """Apply rotary positional encoding to input features.
-
-        Args:
-            feature (Tensor): Input feature tensor.
-            pos_enc (Tensor): Positional encoding tensor.
-            spatial_dim (int): Spatial dimension of input features.
-        """
-
-        assert spatial_dim in (1, 2), f'the argument spatial_dim must be ' \
-            f'either 1 or 2, but got {spatial_dim}'
-
-        for _ in range(feature.ndim - pos_enc.ndim + spatial_dim - 1):
-            pos_enc = pos_enc.unsqueeze(0)
-
-        x1, x2 = torch.chunk(feature, 2, dim=-1)
-        if spatial_dim == 1:
-            cos, sin = torch.chunk(pos_enc, 2, dim=-1)
-            feature = torch.cat((x1 * cos - x2 * sin, x2 * cos + x1 * sin),
-                                dim=-1)
-        elif spatial_dim == 2:
-            pos_enc_h, pos_enc_w = torch.unbind(pos_enc, dim=-1)
-            cos_h, sin_h = torch.chunk(pos_enc_h, 2, dim=-1)
-            cos_w, sin_w = torch.chunk(pos_enc_w, 2, dim=-1)
-            feature = torch.cat(
-                (x1 * cos_h - x2 * sin_h, x1 * cos_w + x2 * sin_w), dim=-1)
-
-        return feature
-
-
-class GAUEncoder(BaseModule):
-    """Gated Attention Unit (GAU) Encoder.
-
-    Args:
-        in_token_dims (int): The input token dimension.
-        out_token_dims (int): The output token dimension.
-        expansion_factor (int, optional): The expansion factor of the
-            intermediate token dimension. Defaults to 2.
-        s (int, optional): The self-attention feature dimension.
-            Defaults to 128.
-        eps (float, optional): The minimum value in clamp. Defaults to 1e-5.
-        dropout_rate (float, optional): The dropout rate. Defaults to 0.0.
-        drop_path (float, optional): The drop path rate. Defaults to 0.0.
-        act_fn (str, optional): The activation function which should be one
-            of the following options:
-
-            - 'ReLU': ReLU activation.
-            - 'SiLU': SiLU activation.
-
-            Defaults to 'SiLU'.
-        bias (bool, optional): Whether to use bias in linear layers.
-            Defaults to False.
-        use_rel_bias (bool, optional): Whether to use relative bias.
-            Defaults to True.
-        pos_enc (bool, optional): Whether to use rotary position
-            embedding. Defaults to False.
-        spatial_dim (int, optional): The spatial dimension of inputs
-
-    Reference:
-        `Transformer Quality in Linear Time
-        <https://arxiv.org/abs/2202.10447>`_
-    """
-
-    def __init__(self,
-                 in_token_dims,
-                 out_token_dims,
-                 expansion_factor=2,
-                 s=128,
-                 eps=1e-5,
-                 dropout_rate=0.,
-                 drop_path=0.,
-                 act_fn='SiLU',
-                 bias=False,
-                 pos_enc=False,
-                 spatial_dim: int = 1):
-
-        super(GAUEncoder, self).__init__()
-        self.s = s
-        self.pos_enc = pos_enc
-        self.spatial_dim = spatial_dim
-        self.drop_path = DropPath(drop_path) \
-            if drop_path > 0. else nn.Identity()
-
-        self.e = int(in_token_dims * expansion_factor)
-        self.o = nn.Linear(self.e, out_token_dims, bias=bias)
-
-        self.uv = nn.Linear(in_token_dims, 2 * self.e + self.s, bias=bias)
-        self.gamma = nn.Parameter(torch.rand((2, self.s)))
-        self.beta = nn.Parameter(torch.rand((2, self.s)))
-
-        self.ln = ScaleNorm(in_token_dims, eps=eps)
-
-        nn.init.xavier_uniform_(self.uv.weight)
-
-        if act_fn == 'SiLU':
-            assert digit_version(TORCH_VERSION) >= digit_version('1.7.0'), \
-                'SiLU activation requires PyTorch version >= 1.7'
-
-            self.act_fn = nn.SiLU(True)
-        else:
-            self.act_fn = nn.ReLU(True)
-
-        if in_token_dims == out_token_dims:
-            self.shortcut = True
-            self.res_scale = Scale(in_token_dims)
-        else:
-            self.shortcut = False
-
-        self.sqrt_s = math.sqrt(s)
-        self.dropout_rate = dropout_rate
-
-        if dropout_rate > 0.:
-            self.dropout = nn.Dropout(dropout_rate)
-
-    def _forward(self, x, mask=None, pos_enc=None):
-        """GAU Forward function."""
-
-        x = self.ln(x)
-
-        # [B, K, in_token_dims] -> [B, K, e + e + s]
-        uv = self.uv(x)
-        uv = self.act_fn(uv)
-
-        # [B, K, e + e + s] -> [B, K, e], [B, K, e], [B, K, s]
-        u, v, base = torch.split(uv, [self.e, self.e, self.s], dim=-1)
-        # [B, K, 1, s] * [1, 1, 2, s] + [2, s] -> [B, K, 2, s]
-        dim = base.ndim - self.gamma.ndim + 1
-        gamma = self.gamma.view(*((1, ) * dim), *self.gamma.size())
-        beta = self.beta.view(*((1, ) * dim), *self.beta.size())
-        base = base.unsqueeze(-2) * gamma + beta
-        # [B, K, 2, s] -> [B, K, s], [B, K, s]
-        q, k = torch.unbind(base, dim=-2)
-
-        if self.pos_enc:
-            q = SinePositionalEncoding.apply_rotary_pos_enc(
-                q, pos_enc, self.spatial_dim)
-            k = SinePositionalEncoding.apply_rotary_pos_enc(
-                k, pos_enc, self.spatial_dim)
-
-        # [B, K, s].transpose(-1, -2) -> [B, s, K]
-        # [B, K, s] x [B, s, K] -> [B, K, K]
-        qk = torch.matmul(q, k.transpose(-1, -2))
-
-        # [B, K, K]
-        kernel = torch.square(F.relu(qk / self.sqrt_s))
-
-        if mask is not None:
-            kernel = kernel * mask
-        if self.dropout_rate > 0.:
-            kernel = self.dropout(kernel)
-
-        # [B, K, K] x [B, K, e] -> [B, K, e]
-        x = u * torch.matmul(kernel, v)
-        # [B, K, e] -> [B, K, out_token_dims]
-        x = self.o(x)
-
-        return x
-
-    def forward(self, x, mask=None, pos_enc=None):
-        """Forward function."""
-        out = self.drop_path(self._forward(x, mask=mask, pos_enc=pos_enc))
-        if self.shortcut:
-            return self.res_scale(x) + out
-        else:
-            return out
